Display flash error messages and Add user's routes
In order to pass variables between pages the connect-flash is used
It allows to streo temporary messages in an area of the session object
called flash. Those messages are cleared once presented to the user.
User's routes for signin, signup and signout are created.
Index view is also updated to greet user and links to options are added